pyproject.toml
[project]
name = "galaxy-sim-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.20.0",
    "moderngl>=5.7.0",
    "glfw>=2.5.5",
    "pyrr>=0.10.3",
    "pillow>=10.4.0",
]
